BookStore Application

+++ src/main/java/com/example/bookstoreapplication/service/BookService.java
@@ -7,7 +7,6 @@ import com.example.bookstoreapplication.model.BookModel;
 import com.example.bookstoreapplication.model.UserModel;
 import com.example.bookstoreapplication.repository.BookRepository;
 import com.example.bookstoreapplication.repository.UserRepository;
-import com.example.bookstoreapplication.utility.EmailService;
 import com.example.bookstoreapplication.utility.JwtUtils;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,9 +29,6 @@ public class BookService implements IbookService{
     ModelMapper modelMapper;
 
     @Autowired
-    EmailService emailService;
-
-    @Autowired
     JwtUtils jwtUtils;
 
 
@@ -41,14 +37,11 @@ public class BookService implements IbookService{
     public BookModel addBooks(String token, BookDTO bookDTO) {
         LoginDTO loginDTO = jwtUtils.decodeToken(token);
         UserModel user = userRepository.findByEmailAndPassword(loginDTO.getEmail(), loginDTO.getPassword());
-        if (userRepository.findByEmail(user.getEmail()).getRole().equals("Admin")) {
-            if (userRepository.findByEmail(user.getEmail()).isLogin()) {
-                BookModel bookModel = modelMapper.map(bookDTO, BookModel.class);
-                return bookRepository.save(bookModel);
-            }
-            throw new BookStoreException("please login Application As admin");
+        if (userRepository.findByEmail(user.getEmail()).getRole().equals("Admin") && userRepository.findByEmail(user.getEmail()).isLogin()) {
+            BookModel bookModel = modelMapper.map(bookDTO, BookModel.class);
+            return bookRepository.save(bookModel);
         }
-        throw new BookStoreException("Only Admin can add Books");
+        throw new BookStoreException("Only Admin can add Books"+"\nplease login Application As admin");
     }
 
     //----------------------------- Update Books Data --------------------------------------------------------------------
@@ -56,37 +49,34 @@ public class BookService implements IbookService{
     public BookModel updateBooksData(String token, int id, BookDTO bookDTO) {
         LoginDTO loginDTO = jwtUtils.decodeToken(token);
         UserModel user = userRepository.findByEmailAndPassword(loginDTO.getEmail(), loginDTO.getPassword());
-        if (userRepository.findByEmail(user.getEmail()).getRole().equals("Admin")) {
-            if (userRepository.findByEmail(user.getEmail()).isLogin()) {
-                if (bookRepository.findById(id).isPresent()) {
-                    BookModel book = bookRepository.findById(id).get();
-                    BookModel updateBook = modelMapper.map(bookDTO, BookModel.class);
-                    updateBook.setBookId(id);
-                    if (updateBook.getBookDescription() == null) {
-                        updateBook.setBookDescription(book.getBookDescription());
-                    }
-                    if (updateBook.getBookName() == null) {
-                        updateBook.setBookName(book.getBookName());
-                    }
-                    if (updateBook.getAuthorName() == null) {
-                        updateBook.setAuthorName(book.getAuthorName());
-                    }
-                    if (updateBook.getBookQuantity() == 0) {
-                        updateBook.setBookQuantity(book.getBookQuantity());
-                    }
-                    if (updateBook.getPrice() == 0) {
-                        updateBook.setPrice(book.getPrice());
-                    }
-                    if (updateBook.getProfilePic() == null) {
-                        updateBook.setProfilePic(book.getProfilePic());
-                    }
-                    return bookRepository.save(updateBook);
+        if (userRepository.findByEmail(user.getEmail()).getRole().equals("Admin") && userRepository.findByEmail(user.getEmail()).isLogin()) {
+            if (bookRepository.findById(id).isPresent()) {
+                BookModel book = bookRepository.findById(id).get();
+                BookModel updateBook = modelMapper.map(bookDTO, BookModel.class);
+                updateBook.setBookId(id);
+                if (updateBook.getBookDescription() == null) {
+                    updateBook.setBookDescription(book.getBookDescription());
                 }
-                throw new BookStoreException("Invalid Id");
+                if (updateBook.getBookName() == null) {
+                    updateBook.setBookName(book.getBookName());
+                }
+                if (updateBook.getAuthorName() == null) {
+                    updateBook.setAuthorName(book.getAuthorName());
+                }
+                if (updateBook.getBookQuantity() == 0) {
+                    updateBook.setBookQuantity(book.getBookQuantity());
+                }
+                if (updateBook.getPrice() == 0) {
+                    updateBook.setPrice(book.getPrice());
+                }
+                if (updateBook.getProfilePic() == null) {
+                    updateBook.setProfilePic(book.getProfilePic());
+                }
+                return bookRepository.save(updateBook);
             }
-            throw new BookStoreException("please login Application As admin");
+            throw new BookStoreException("Invalid Id");
         }
-        throw new BookStoreException("Only Admin can Update Books");
+        throw new BookStoreException("Only Admin can Update Books"+"\nplease login Application As admin");
     }
 
     //--------------------------------- Delete Book Data ------------------------------------------------------------------
@@ -94,17 +84,14 @@ public class BookService implements IbookService{
     public String deleteBookById(String token, int id) {
         LoginDTO loginDTO = jwtUtils.decodeToken(token);
         UserModel user = userRepository.findByEmailAndPassword(loginDTO.getEmail(), loginDTO.getPassword());
-        if (userRepository.findByEmail(user.getEmail()).getRole().equals("Admin")) {
-            if (userRepository.findByEmail(user.getEmail()).isLogin()) {
-                if (bookRepository.findById(id).isPresent()) {
-                    bookRepository.deleteById(id);
-                    return "Book Data Deleted successful";
-                }
-                throw new BookStoreException("Book Not Found "+"\nInvalid Id ");
+        if (userRepository.findByEmail(user.getEmail()).getRole().equals("Admin") && userRepository.findByEmail(user.getEmail()).isLogin()) {
+            if (bookRepository.findById(id).isPresent()) {
+                bookRepository.deleteById(id);
+                return "Book Data Deleted successful";
             }
-            throw new BookStoreException("please login Application As admin");
+            throw new BookStoreException("Book Not Found "+"\nInvalid Id ");
         }
-        throw new BookStoreException("Only Admin can Delete Books");
+        throw new BookStoreException("Only Admin can Delete Books"+"\nplease login Application As admin");
     }
 
     //----------------------------- Show All Books Data --------------------------------------------------------------------
@@ -134,6 +121,16 @@ public class BookService implements IbookService{
             throw new BookStoreException("Book with name " + bookName + " is not found!");
         }
         return bookList;
+    }
+
+    //----------------------------- Search Books by Author Name --------------------------------------------------------------------
+    @Override
+    public List<BookModel> searchBookByAuthorName(String authorName) {
+        List<BookModel> bookModelList = bookRepository.findAllByAuthorName(authorName);
+        if (bookModelList.isEmpty()) {
+            throw new BookStoreException("Book with Author Name " + authorName + " is not found!");
+        }
+        return bookModelList;
     }
 
     //--------------------------------- Sort Book Data By Price High To Low ---------------------------------
